backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt==3.2.2",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.124.4",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "pyjwt>=2.10.1",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "slowapi>=0.1.9",
]
